Changes to documentation after review
The core review comment was that it was not clear:
* what the upgrade/downgrade doc was for
* how it should be used

To that end a new Introduction/Purpose/Scope/Usage has been written.

The document has also been renamed.

In addition a number of specific points have been fixed.

It has become apparent from other documentation PRs that all
doco's should be spell checked - so that has been done with
a predictable improvement in spelling