Major clean up to address @russelldb pointing out an error in hows keys are laid out on disk.
In the previous version I had an erroronous thing called 'keyspace'.

For a bucket that has an n_val of 3 a vnode contains keys that hash to it,
and to the previous bucket and the previous but one. These keys are intermingled
on disk and no seperated into the spurious 'keyspace'.

On each diagram of writes at the vnode the diagram had to be edited to remove that.